% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.simulation.cov.R
\name{application.covariates}
\alias{application.covariates}
\title{Simulation of applying different dimensions of covariates}
\usage{
application.covariates(model)
}
\arguments{
\item{model}{choose from "IV" and "CTE"}
}
\value{
files with certain name tags, e.g. "linear.cov.cte.Rdata"
}
\description{
Simulation of applying different dimensions of covariates
}
