% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.simulation.lasso.R
\name{application.lasso}
\alias{application.lasso}
\title{Simulation of using lasso for linear and nonlinear structure model}
\usage{
application.lasso(model)
}
\arguments{
\item{model}{choose from "IV" and "CTE"}
}
\value{
files with certain name tags, e.g. "linear.lasso.cte.Rdata"
}
\description{
Simulation of using lasso for linear and nonlinear structure model
}
