% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.simulation.samples.R
\name{application.samples}
\alias{application.samples}
\title{Simulation of applying different number of samples}
\usage{
application.samples(model)
}
\arguments{
\item{model}{choose from "IV" and "CTE"}
}
\value{
files with certain name tags, e.g. "linear.sample.cte.Rdata"
}
\description{
Simulation of applying different number of samples
}
