% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.simulation.sd.R
\name{application.sd}
\alias{application.sd}
\title{Simulation of using different standard errors in treatment generation}
\usage{
application.sd()
}
\value{
files with certain name tags, e.g. "polynom.sd1.Rdata"
}
\description{
Simulation of using different standard errors in treatment generation
}
