% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.evaluation.R
\name{bias.evaluation}
\alias{bias.evaluation}
\title{Bias Evaluation}
\usage{
bias.evaluation(result.list, true.data)
}
\arguments{
\item{result.list}{result list from simulation.wrap function}

\item{true.data}{result from true.data.generation function}
}
\value{
bias as a numeric
}
\description{
Bias Evaluation
}
