% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.evaluation.R
\name{bias.evaluation.wrap}
\alias{bias.evaluation.wrap}
\title{Bias Evaluation Wrap-up}
\usage{
bias.evaluation.wrap(res.list.list, true.data)
}
\arguments{
\item{true.data}{result from true.data.generation function}

\item{result.list.list}{list of result lists from running.simulation function}
}
\value{
a list of bias lists
}
\description{
Bias Evaluation Wrap-up
}
