% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/generations.of.covariates,noises,treatments,responses.R
\name{covariates.generation}
\alias{covariates.generation}
\title{Generation of the covariates dataset}
\usage{
covariates.generation(distributionOption = "normal", numOfCovariates = 2,
  numOfSamples)
}
\arguments{
\item{distributionOption}{choose from normal, uniform, student distributions}

\item{numOfCovariates}{dimension of covariates}

\item{numOfSamples}{number of samples}
}
\value{
A matrix the columns of which represent different covariates, the rows of which represents different observations.
}
\description{
Generation of the covariates dataset
}
