% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.generation.R
\name{data.generation}
\alias{data.generation}
\title{Data Generation}
\usage{
data.generation(model = "CTE", numOfSamples = 1000,
  numOfPreSamples = NULL, numOfCovariates = 2,
  distributionForCovariates = "normal", distributionForIVs = "normal",
  distributionForNoise = "normal", distributionForTreatment = "normal",
  parametricCurveOption = "linear", trimLowerBound = -10,
  trimUpperBound = 10, sd = 8)
}
\arguments{
\item{model}{choose from "CTE", "IV"}

\item{numOfSamples}{numer of samples, default = 1000.}

\item{numOfPreSamples}{number of samples before sample trimming. As default, we set this as NULL}

\item{numOfCovariates}{dimension of covariates, default = 2.}

\item{distributionForCovariates}{choose from normal, uniform, student distributions.}

\item{distributionForIVs}{choose from normal, uniform, student distributions.}

\item{distributionForNoise}{choose from normal, uniform, student distributions.}

\item{distributionForTreatment}{choose from normal, uniform, student distributions.}

\item{parametricCurveOption}{choose from "linear", "polynom", "polynom2", "polynom3", "mixture".}

\item{trimLowerBound}{the lower bound of treatment trimming t.}

\item{trimUpperBound}{the upper bound of treatment trimming t.}

\item{sd}{\eqn{t = \sum x_i + \epsilon}, the standard error of \eqn{\epsilon}}
}
\value{
frame of generated data.
}
\description{
Data Generation
}
\examples{
data.generation(model = "CTE", numOfSamples = 1000,
numOfCovariates = 2,
distributionForCovariates = "normal",
distributionForNoise = "normal",
distributionForTreatment = "normal",
parametricCurveOption = "linear",
trimLowerBound = -10,
trimUpperBound = 10)
}
