% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.generation.R
\name{data.generation.wrap}
\alias{data.generation.wrap}
\title{Data Generation wrap-up (short-cut name: dg.w)}
\usage{
data.generation.wrap(simu = 100, samples = 1000, covariates = 2, model,
  trimLowerBound = -4, trimUpperBound = 4,
  parametricCurveOption = "polynom", ncores = ncores, sd = 8)
}
\arguments{
\item{simu}{number of times one would run, default = 100.}

\item{samples}{number of samples, default = 1000}

\item{covariates}{dimension of covariates}

\item{model}{choose from "CTE", "IV"}

\item{trimLowerBound}{the lower bound of treatment trimming t, default = -4.}

\item{trimUpperBound}{the lower bound of treatment trimming t, default = 4.}

\item{parametricCurveOption}{choose from "linear", "polynom", "polynom2", "polynom3", "mixture".}

\item{ncores}{number of cores we use}

\item{sd}{\eqn{t = \sum x_i + \epsilon}, the standard error of \eqn{\epsilon}, choose from 1, 2, 3, 5, 8, 10, 15}
}
\value{
list that contains two parts:
                $data contains generated data,
                $true contains the true underlying response and the density at grid on treatment
}
\description{
Data Generation wrap-up (short-cut name: dg.w)
}
