% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{default_learners}
\alias{default_learners}
\title{Default Learners List}
\usage{
default_learners()
}
\value{
A list of instantiated learners from various algorithms.
}
\description{
This list defines a set of default learners for use in machine learning tasks.
The learners include generalized linear models, multivariate adaptive regression splines,
generalized additive models, random forests, and boosted trees with early stopping.
}
