% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.evaluation.R
\name{error.evaluation.wrap}
\alias{error.evaluation.wrap}
\title{Error Evaluation Wrap-up}
\usage{
error.evaluation.wrap(true.data, result.list, ncores)
}
\arguments{
\item{true.data}{result from true.data.generation function}

\item{result.list}{result list from simulation.wrap function}
}
\value{
a list of error lists where "error1" stands for l^1 error, "error2" stands for "l^2" error
}
\description{
Error Evaluation Wrap-up
}
