% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.evaluation.R
\name{error1.evaluation}
\alias{error1.evaluation}
\title{l^1 error evaluation}
\usage{
error1.evaluation(result, true.data)
}
\arguments{
\item{result}{result from simulation function}

\item{true.data}{result from true.data.generation function}
}
\value{
L1 error
}
\description{
l^1 error evaluation
}
