% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.evaluation.R
\name{error2.evaluation}
\alias{error2.evaluation}
\title{l^2 error evaluation}
\usage{
error2.evaluation(result, true.data)
}
\arguments{
\item{result}{result from simulation function}

\item{true.data}{result from true.data.generation function}
}
\value{
L2 error
}
\description{
l^2 error evaluation
}
