% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdml.R
\name{estimate_cdml}
\alias{estimate_cdml}
\title{Estimate Calibrated Debiased Machine Learning (CDML)}
\usage{
estimate_cdml(
  A,
  Y,
  mu1,
  mu0,
  pi1,
  pi0,
  weights = NULL,
  functional = NULL,
  representer = NULL
)
}
\arguments{
\item{A}{A binary indicator variable (1 for treatment, 0 for control).}

\item{Y}{Observed outcomes.}

\item{mu1}{Cross-fitted outcome regression estimates for the treated group.}

\item{mu0}{Cross-fitted outcome regression estimates for the control group.}

\item{pi1}{Cross-fitted propensity score estimates for the treatment group (\code{A = 1}).}

\item{pi0}{Cross-fitted propensity score estimates for the control group (\code{A = 0}).}

\item{weights}{Optional weights to use in the estimation. Defaults to equal weights if \code{NULL}.}

\item{functional}{Function to compute the desired functional of interest, given relevant inputs.}

\item{representer}{Function to compute the debiasing term for the estimator.}
}
\value{
The estimated causal effect using the CDML estimator.
}
\description{
Extends the DML estimate by calibrating outcome regression estimates and propensity scores.
Implements the calibrated version of the CDML estimator for causal inference.
}
\details{
This function calibrates the outcome regression and inverse propensity weights before
using the CDML framework for estimation. Calibration improves estimation accuracy and
robustness, leveraging cross-fitting and doubly robust methods.
}
\seealso{
\code{\link{estimate_dml}}, \code{\link{calibrate_outcome_regression}}, \code{\link{calibrate_inverse_weights}}
}
