% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdml.R
\name{estimate_dml}
\alias{estimate_dml}
\title{Compute Calibrated Debiased Machine Learning Estimate (ICDML)}
\usage{
estimate_dml(
  A,
  Y,
  mu1,
  mu0,
  pi1,
  pi0,
  weights = NULL,
  functional = NULL,
  representer = NULL
)
}
\arguments{
\item{A}{A binary indicator variable (1 for treatment, 0 for control).}

\item{Y}{Observed outcomes.}

\item{mu1}{Cross-fitted outcome regression estimates for the treated group.}

\item{mu0}{Cross-fitted outcome regression estimates for the control group.}

\item{pi1}{Cross-fitted propensity score estimates for the treatment group (\code{A = 1}).}

\item{pi0}{Cross-fitted propensity score estimates for the control group (\code{A = 0}).}

\item{weights}{Optional weights to use in the estimation. Defaults to equal weights if \code{NULL}.}

\item{functional}{Function to compute the desired functional of interest, given relevant inputs.}

\item{representer}{Function to compute the debiasing term for the estimator.}
}
\value{
A numeric value representing the estimated causal effect using the ICDML estimator.
}
\description{
Calculates the Calibrated Debiased Machine Learning (C-DML) estimator for
causal inference, specifically for estimating linear functionals of the outcome regression.
}
\details{
The function computes a C-DML estimate, which involves estimating a "plug-in" term and a
debiasing term based on cross-fitted outcome regressions and propensity scores. This function
incorporates doubly robust properties to mitigate estimation bias and improve robustness.
}
\seealso{
\code{\link{estimate_cdml}}, \code{\link{calibrate_outcome_regression}}, \code{\link{calibrate_inverse_weights}}
}
