% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_DML.R
\name{estimate_IC_DML}
\alias{estimate_IC_DML}
\title{Compute Calibrated Debiased Machine Learning Estimate (ICDML)}
\usage{
estimate_IC_DML(A, Y, mu1, mu0, pi1, pi0)
}
\arguments{
\item{A}{A binary indicator variable (1 for treatment, 0 for control).}

\item{Y}{Observed outcomes.}

\item{mu1}{Cross-fitted outcome regression estimates for the treated group.}

\item{mu0}{Cross-fitted outcome regression estimates for the control group.}

\item{pi1}{Cross-fitted propensity score estimates for the treatment group (\code{A = 1}).}

\item{pi0}{Cross-fitted propensity score estimates for the control group (\code{A = 0}).}
}
\value{
A numeric value representing the estimated causal effect, \eqn{\tau_n}, using the ICDML estimator.
}
\description{
Calculates the Calibrated Debiased Machine Learning (C-DML) estimator for
causal inference, specifically for estimating linear functionals of the outcome regression.
}
\details{
This function implements the ICDML estimator, a novel calibrated debiased machine learning
approach designed for estimating causal effects in a doubly robust manner.
Doubly robust consistency and inference are achieved through a combination of cross-fitting,
isotonic calibration, and debiased machine learning.

The ICDML estimator is part of the C-DML framework, which leverages the link between calibration
and doubly robust asymptotic linearity. By integrating outcome regression calibration
and inverse propensity weights, ICDML maintains asymptotic linearity when either the
outcome regression or the Riesz representer is sufficiently well-estimated, even if the other
is inconsistently or slowly estimated. This robustness enables valid inference with flexible
machine learning models.

For confidence intervals, a bootstrap-assisted approach is recommended to maintain doubly
robust inference properties.
}
\seealso{
\code{\link{calibrate_outcome_regression}}, \code{\link{calibrate_inverse_weights}}
}
