% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.estimation.R
\name{g.method.estimation}
\alias{g.method.estimation}
\title{Regression Method Estimation}
\usage{
g.method.estimation(data, g.method)
}
\arguments{
\item{data}{data that we based estimation on.}

\item{g.method}{choose regression method from "nnet", "trees", "rf", "MARS", "lasso".}
}
\value{
results of correspondent method estimation
}
\description{
Regression Method Estimation
}
