% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.estimation.R
\name{g.predict}
\alias{g.predict}
\title{Regression Method Prediction}
\usage{
g.predict(object, data.eval, data.chunk, mode = "CDML", verbose = TRUE,
  g.method = "rf")
}
\arguments{
\item{object}{object that is from g.method.estimation function}

\item{data.eval}{data to be evaluated, the (T,X) pair is of interest}

\item{data.chunk}{whole data, covariates are of interest}

\item{mode}{choose from "CDML", "SR", "HI". The default set-up is for "CDML" case.}

\item{verbose}{whether to print step in console}

\item{g.method}{regression g.method that is specified in parameter object.}
}
\value{
a vector and a matrix
For the matrix elements, they stand for the predicted values of:
row means given certain treatment from data.eval
column means given certain covariates from data.chunk.
For the vector: we list in order the predicted value as \eqn{i=1,...,n} for \eqn{(T_i, X_i)}
}
\description{
Regression Method Prediction
}
