% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.predict.R
\name{gps.predict}
\alias{gps.predict}
\title{Generalize Propensity Score Prediction}
\usage{
gps.predict(object, gps.method = "series", data.chunk = NULL, data.eval,
  t.grid = NULL, grid.length = 1000, verbose = FALSE,
  treatment.min = -4, treatment.max = 4, detoured = FALSE)
}
\arguments{
\item{object}{an object from gps.method.estimation function}

\item{gps.method}{selected generalized propensity score method}

\item{data.chunk}{whole data, covariates are of interest. defualt = NULL}

\item{data.eval}{data to be evaluated, the treatment is of interest}

\item{t.grid}{alternate parameter to set up if data.chunk is NULL}

\item{grid.length}{how dense the evaluation point should be like}

\item{verbose}{print in console the procedure step by step}

\item{detoured}{if FALSE method will be done normally;
otherwise a special simulation will be examined, where CDML is chosen method, CTE is chosen model,
gps.method "series" is not allowed.}
}
\value{
a vector and a matrix
For the matrix elements, they stand for the predicted values of:
row means given certain treatment from data.eval
column means given certain covariates from data.chunk.
For the vector: we list in order the predicted value as \eqn{i=1,...,n} for \eqn{(T_i, X_i)}
}
\description{
Generalize Propensity Score Prediction
}
