% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/generations.of.covariates,noises,treatments,responses.R
\name{instrument.generation}
\alias{instrument.generation}
\title{Generation of instrument variable dataset}
\usage{
instrument.generation(distributionOption = "normal", covariates, numOfSamples)
}
\arguments{
\item{distributionOption}{choose from normal, uniform, student distributions}

\item{covariates}{dimension of covariates}

\item{numOfSamples}{number of samples}
}
\value{
a vector of instrumental variable
}
\description{
Generation of instrument variable dataset
}
