% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.alternative.R
\name{plot.alternative.inside}
\alias{plot.alternative.inside}
\title{Plot of Alternatives}
\usage{
\method{plot}{alternative.inside}(gps.method, g.method,
  method.involved = c("CDML.1", "CDML.2"), sample.size.option = NULL,
  num.cov.option = c(2, 5), treatment.sd.option = NULL,
  struct.model.option = NULL, what, file = NULL, plot.what = "bias",
  main = TRUE, color = "red")
}
\arguments{
\item{gps.method}{a vector of method for generalized propensity score estimation}

\item{g.method}{a vector of method for regression estimation}

\item{method.involved}{e.g.: c("SR", "CDML.1", "CDML.2", "CDML.5", "HI")}

\item{sample.size.option}{a vector of number of samples that will be selected from}

\item{num.cov.option}{a vector of number of covariates that will be selected from}

\item{treatment.sd.option}{a vector of standard deviation of treatment generation that will be selected from}

\item{struct.model.option}{a vector of parametric curves of response generation, typically choose from "linear", "polynom", "polynom2", "polynom3", "mixture".}

\item{what}{choose from "sample.sizes", "cov"}

\item{file}{file that we save our return into}

\item{plot.what}{choose from "bias" and "error"}

\item{main}{if main == TRUE, we plot a new one, otherwise we attach our figure in an existing plot}

\item{color}{can be used to specify the color of the main line}
}
\value{
plots
}
\description{
Plot of Alternatives
}
