% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/generations.of.covariates,noises,treatments,responses.R
\name{response.generation}
\alias{response.generation}
\title{Generation of response dataset}
\usage{
response.generation(noise, covariates, treatment,
  parametricCurveOption = "linear", numOfSamples = 1000)
}
\arguments{
\item{noise}{the noise vector generated by function noise.generation(), term u in the model}

\item{covariates}{the covariates matrix generated by function covariates.generation(), term x in the model}

\item{treatment}{the treatment vector generated by function treatment.generation(), term t in the model}

\item{parametricCurveOption}{g(x,t) with different smoothness chosen from "linear", "polynom", "polynom2", "polynom3", "mixture"}

\item{numOfSamples}{number of samples}
}
\value{
a vector of responses
}
\description{
Generation of response dataset
}
