% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running.simulation.R
\name{running.simulation}
\alias{running.simulation}
\title{Running Simulation}
\usage{
running.simulation(model = "CTE", simu = 40, samples = c(100, 200, 500),
  g.method = c("nnet", "rf", "MARS"), gps.method = c("series",
  "boosting&normal", "linear&boxcox"), trimming = c(-4, 4), cov = c(2, 5,
  10), method = c("SR", "HI", "CDML"), fold = c(1, 2, 3, 5),
  responseCurve = "polynom", file = "r_scratch/cte100and200and500.Rdata",
  sd = 8)
}
\arguments{
\item{model}{model varies from "IV", "CTE"}

\item{simu}{number of simulations one would run, default = 100.}

\item{samples}{number of samples}

\item{g.method}{a vector of method for regression estimation}

\item{gps.method}{a vector of method for generalized propensity score estimation}

\item{trimming}{trimming of treatment vector, default = c(-4, 4)}

\item{cov}{a vector of dimensions of covariates}

\item{method}{a vector of estimation methods we choose from double machine learning method ("CDML")
and simple regression method ("SR"), Hirano & Imbens method ("HI")}

\item{fold}{number of folds for sample splitting}

\item{responseCurve}{choose from "linear", "polynom", "polynom2", "polynom3", "mixture".}

\item{file}{file that we save our return into}

\item{sd}{\eqn{t = \sum x_i + \epsilon}, the standard error of \eqn{\epsilon}, choose from 1, 2, 3, 5, 8, 10, 15}
}
\value{
no returns
}
\description{
Running Simulation
}
\examples{
running.simulation(
########### enter all parameters needed for running simulations ##########
model <- "CTE",
simu <-8,
samples <- c(200),
g.method <- c("lasso"),
gps.method <-
  c(#"series",
   "rf&normal" ,
   "linear&boxcox"),
trimming <- c(-4, 4),
cov <- c(5),
method <- c("SR", "CDML"),
fold <- c(1, 2, 3),
responseCurve <- "polynom3",
file <- "demo.Rdata", # the file saves data
sd =8)
}
