% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.alternative.R
\name{simulation.alternative}
\alias{simulation.alternative}
\title{The alternative of running simulation}
\usage{
simulation.alternative()
}
\value{
file saved with name specified
}
\description{
The alternative of running simulation
}
