% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationBodyPart.R
\name{simulation}
\alias{simulation}
\title{Simulation Body Part}
\usage{
simulation(kfold, data = data, trimLowerBound.pi = 1e-05,
  trimLowerBound.pibar = 1e-05, trimUpperBound.pi = 10,
  trimUpperBound.pibar = 10, method = "CDML", g.method = "rf",
  gps.method = "series", trimLowerBound.t = -10, trimUpperBound.t = 10,
  model, variation = "whole", verbose = TRUE)
}
\arguments{
\item{kfold}{number of folds for sample splitting}

\item{data}{data}

\item{trimLowerBound.pi}{trimming lower bound of pi estimation}

\item{trimLowerBound.pibar}{trimming lower bound of pibar estimation}

\item{trimUpperBound.pi}{trimming upper bound of pi estimation}

\item{trimUpperBound.pibar}{trimming upper bound of pibar estimation}

\item{method}{estimation method we choose from double machine learning method ("CDML")
and simple regression method ("SR"), Hirano & Imbens method ("HI")}

\item{g.method}{method for regression estimation}

\item{gps.method}{method for generalized propensity score estimation}

\item{trimLowerBound.t}{trimming lower bound of treatment vector}

\item{trimUpperBound.t}{trimming upper bound of treatment vector}

\item{model}{model varies from "IV", "CTE"}

\item{variation}{choose from "whole","train.data".
Estimate the probability measure by empirical measure
                  of covariates. "whole" uses every data sample,
                   "train.data" only trainning data for estimating empirical measure.
                   As default, we select "whole"}

\item{verbose}{print in console step by step}
}
\value{
a list of estimated value "yhat" evaluated at "t.grid", with "record" to record details of this simulation
}
\description{
Simulation Body Part
}
