% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.wrap.R
\name{simulation.wrap}
\alias{simulation.wrap}
\title{Simulation wrap-up (short-cut name: sim.w)}
\usage{
simulation.wrap(kfold, data, model, ncores = ncores, method,
  g.method = "rf", gps.method = "series", trimLowerBound.t = -4,
  trimUpperBound.t = 4)
}
\arguments{
\item{kfold}{number of folds for sample splitting}

\item{data}{data}

\item{model}{model varies from "IV", "CTE"}

\item{ncores}{number of cores we use}

\item{method}{estimation method we choose from double machine learning method ("CDML")
and simple regression method ("SR"), Hirano & Imbens method ("HI")}

\item{g.method}{method for regression estimation}

\item{gps.method}{method for generalized propensity score estimation}

\item{trimLowerBound.t}{trimming lower bound of treatment vector}

\item{trimUpperBound.t}{trimming upper bound of treatment vector}
}
\value{
a list of returned objects in function simulation.
}
\description{
Simulation wrap-up (short-cut name: sim.w)
}
