% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/generations.of.covariates,noises,treatments,responses.R
\name{treatment.generation}
\alias{treatment.generation}
\title{Generation of treatment dataset}
\usage{
treatment.generation(distributionOption = "normal", numOfSamples = 1000,
  covariates, sd = 8)
}
\arguments{
\item{distributionOption}{choose from normal, uniform, student distributions}

\item{numOfSamples}{number of samples}

\item{covariates}{the object that returned from function covariates.generation}

\item{sd}{\eqn{t = \sum x_i + \epsilon}, the standard error of \eqn{\epsilon}}
}
\value{
A vector stands for treatment
}
\description{
Generation of treatment dataset
}
