% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true.data.generation.R
\name{true.data.generation}
\alias{true.data.generation}
\title{True Data Generation}
\usage{
true.data.generation(numOfCovariates = 2,
  distributionForCovariates = "normal", distributionForNoise = "normal",
  distributionForTreatment = "normal", parametricCurveOption = "polynom",
  trimLowerBound = -4, trimUpperBound = 4, nrep = 1e+06)
}
\arguments{
\item{numOfCovariates}{dimension of covariates}

\item{distributionForCovariates}{default = "normal", choose from normal, student, uniform distributions}

\item{distributionForNoise}{default = "normal", choose from normal, student, uniform distributions}

\item{distributionForTreatment}{default = "normal", choose from normal, student, uniform distributions}

\item{parametricCurveOption}{default = "linear", choose from "linear", "polynom", "polynom2", "polynom3", "mixture"}

\item{trimLowerBound}{the lower bound of trimming treatment vector}

\item{trimUpperBound}{the upper bound of trimming treatment vector}

\item{nrep}{default = 1000000, numerically calculate the true density by setting equi-distant nrep sample points for approximation}
}
\value{
a list of true response value "true.y" evaluated at "grid", with treatment density "density".
}
\description{
True Data Generation
}
